API

ComplexValues.Polar — Method

Polar(z)

Construct a polar representation of z.

ComplexValues.Polar — Method

Polar(radius::Real, azimuth::Real)

Construct a polar value with the given radius and angle.

ComplexValues.Spherical — Method

Spherical(z)

Construct a spherical representation of the value z.

ComplexValues.Spherical — Method

Spherical(latitude, azimuth)

Construct a spherical representation with given latitude in [-π/2,π/2] and azimuth.

ComplexValues.S2coord — Method

S2coord(u::Spherical)

Convert the spherical value to a 3-vector of coordinates on the unit sphere.

ComplexValues.real_type — Method

real_type(z)

Return the base type of a real or complex float z, i.e., the type of its real part.

Examples

julia> real_type(1.0)
Float64
julia> real_type(1.0f0 + 2im)
Float32
julia> real_type(Polar(BigFloat(1), 2))
BigFloat