Spherical

Spherical

A Spherical value stores the latitude and azimuthal angles. For efficiency, the angles are not converted to any standard interval, though some operations on values do so. They can be accessed directly as the lat and lon (for longitude) fields.

Multiplicative inverse (inv) and sign of Spherical values exploit the structure of the representation. Otherwise, most operations are done by first converting to Complex.

The comparators iszero, isinf, isfinite, and isapprox are defined appropriately.

Using Plots, a plot of Spherical values will be drawn on the surface of the unit sphere. The keyword sphere is used to control the additional plotting of a gray wireframe on the sphere. If set to true (the default), the number of latitude and longitude circles is chosen automatically; if set to a tuple, it determines the number of longitudinal and latitudinal circles, respectively.

ComplexValues.Spherical — Type.

(type) Spherical

Representation of a complex value on the Riemann sphere.

source

ComplexValues.Spherical — Method.

Spherical(z) Construct a spherical representation of the value z.

source

ComplexValues.Spherical — Method.

Spherical(latitude,azimuth)

Construct a spherical representation with given latitude in [-π/2,π/2] and azimuth.

source

ComplexValues.S2coord — Method.

S2coord(u::Spherical)

Convert the spherical value to a 3-vector of coordinates on the unit sphere.

source